Makefile portability, dependency graph and error handling updates.

+++ Makefile
@@ -2,10 +2,11 @@ SHELL = /bin/sh
 
 .SILENT:
 
-.PHONY: build check analyze clean
+.PHONY: dirs build check analyze clean
 
 LIB_NAME := ctap
 SRC_DIR := src
+TEST_DIR := test
 BUILD_DIR := build
 UTEST_DIR := test/unit
 ITEST_DIR := test/integration
@@ -13,7 +14,7 @@ UTILTEST_DIR := test/utilities
 
 CC := clang
 AR := ar rcs 
-CFLAGS := -std=c99 -g -I. 
+CFLAGS := -std=c99 -g -I. -MMD
 RELEASE_CFLAGS := -O3 -flto -finline-functions \
 				  -ffunction-sections -fdata-sections \
 				  -ffreestanding \
@@ -29,10 +30,11 @@ DEBUG_CFLAGS := -DDEBUG -O0 -Weverything -Werror -fsanitize=address \
 				$(CFLAGS)
 
 SRCS := $(wildcard $(SRC_DIR)/*.c)
+OBJS := $(SRCS:%.c=$(BUILD_DIR)/%.o)
+DEPS = $(OBJS:.o=.d)
 ITESTS := $(wildcard $(ITEST_DIR)/*.c)
 UTESTS := $(wildcard $(UTEST_DIR)/*.c)
 UTILTESTS := $(wildcard $(UTILTEST_DIR)/*.c)
-OBJS := $(SRCS:%.c=$(BUILD_DIR)/%.o)
 UTEST_BINS := $(UTESTS:%.c=$(BUILD_DIR)/%)
 ITEST_BINS := $(ITESTS:%.c=$(BUILD_DIR)/%)
 UTILTEST_BINS := $(UTILTESTS:%.c=$(BUILD_DIR)/%)
@@ -40,24 +42,25 @@ UTEST_LOGS := $(patsubst %,%.log,$(UTEST_BINS))
 ITEST_LOGS := $(patsubst %,%.log,$(ITEST_BINS))
 UTILTEST_LOGS := $(patsubst %,%.log,$(UTILTEST_BINS))
 
+# Include header dependency targets if they've been generated with -MMD.
+-include $(DEPS)
+
 build: CFLAGS = $(RELEASE_CFLAGS)
-build: $(BUILD_DIR)/$(LIB_NAME).a
+build: dirs $(BUILD_DIR)/$(LIB_NAME).a
 
 check: CFLAGS = $(RELEASE_CFLAGS)
-check: $(BUILD_DIR)/$(UTILTEST_DIR) $(BUILD_DIR)/$(ITEST_DIR) $(BUILD_DIR)/$(UTEST_DIR) \
-	   $(UTEST_LOGS) $(ITEST_LOGS) $(UTILTEST_LOGS) 
+check: dirs $(UTEST_LOGS) $(ITEST_LOGS) $(UTILTEST_LOGS)
 
-$(BUILD_DIR)/$(SRC_DIR):
-	mkdir -p $(BUILD_DIR)/$(SRC_DIR)
-$(BUILD_DIR)/$(UTEST_DIR):
-	mkdir -p $(BUILD_DIR)/$(UTEST_DIR)
-$(BUILD_DIR)/$(ITEST_DIR):
-	mkdir -p $(BUILD_DIR)/$(ITEST_DIR)
-$(BUILD_DIR)/$(UTILTEST_DIR):
-	mkdir -p $(BUILD_DIR)/$(UTILTEST_DIR)
+dirs:
+	test -d $(BUILD_DIR) || mkdir $(BUILD_DIR)
+	test -d $(BUILD_DIR)/$(SRC_DIR) || mkdir $(BUILD_DIR)/$(SRC_DIR)
+	test -d $(BUILD_DIR)/$(TEST_DIR) || mkdir $(BUILD_DIR)/$(TEST_DIR)
+	test -d $(BUILD_DIR)/$(UTEST_DIR) || mkdir $(BUILD_DIR)/$(UTEST_DIR)
+	test -d $(BUILD_DIR)/$(ITEST_DIR) || mkdir $(BUILD_DIR)/$(ITEST_DIR)
+	test -d $(BUILD_DIR)/$(UTILTEST_DIR) || mkdir $(BUILD_DIR)/$(UTILTEST_DIR)
 
 # Create a static library archive.
-$(BUILD_DIR)/$(LIB_NAME).a: $(BUILD_DIR)/$(SRC_DIR) $(OBJS) 
+$(BUILD_DIR)/$(LIB_NAME).a: $(OBJS) 
 	$(AR) $(BUILD_DIR)/$(LIB_NAME).a $(OBJS)
 
 # Build object files.
@@ -89,4 +92,4 @@ $(BUILD_DIR)/$(UTEST_DIR)/%: $(UTEST_DIR)/%.c
 	$(CC) $< $(patsubst $(UTEST_DIR)/%_test.c,$(SRC_DIR)/%*.c,$<) -o $@ $(CFLAGS)
 
 clean:
-	-rm -r $(BUILD_DIR)
+	-test -d $(BUILD_DIR) && rm -r $(BUILD_DIR)
